data/tasktracker/requirements.txt
amqp==5.0.5
asgiref==3.3.1
astroid==2.4.2
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
clickhouse-driver==0.2.0
colorama==0.4.4
Django==3.1.6
django-crispy-forms==1.11.0
djangorestframework==3.12.2
isort==5.7.0
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
prompt-toolkit==3.0.14
pylint==2.6.0
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1